# tools/dashboard/backend/docs/swagger.yaml
definitions:
...
  xstores.ContainerStatusDTO:
    description: Status of a container within a pod.
    properties:
      image:
        description: Container image
        example: polardbx/polardbx-engine:latest
        type: string
      name:
        description: Container name
        example: engine
        type: string
      ready:
        description: Whether container is ready
        example: true
        type: boolean
      restartCount:
        description: Restart count
        example: 0
        type: integer
    type: object
...
  xstores.XStorePodDTO:
    description: Pod information for XStore instance.
    properties:
      containerStatuses:
        description: Container statuses
        items:
          $ref: '#/definitions/xstores.ContainerStatusDTO'
        type: array
      creationTimestamp:
        description: Pod creation timestamp
        type: string
      hostIP:
        description: Host IP where pod is running
        example: 192.168.1.100
        type: string
      name:
        description: Pod name
        example: my-xstore-0
        type: string
      namespace:
        description: Pod namespace
        example: default
        type: string
      nodeName:
        description: Node name where pod is scheduled
        example: worker-node-1
        type: string
      phase:
        description: Pod phase (Pending, Running, Succeeded, Failed, Unknown)
        example: Running
        type: string
      podIP:
        description: Pod IP address
        example: 10.244.0.15
        type: string
      role:
        description: Pod role (Leader, Follower, Learner)
        example: Leader
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/platform/logs/{namespace}/{pod_name}:
    get:
      consumes:
      - application/json
      description: Retrieves logs from a specific pod, optionally filtered by container
      parameters:
      - description: 'Kubernetes namespace (default: default)'
        in: query
        name: namespace
        type: string
      - description: Name of the pod
        in: path
        name: pod_name
        required: true
        type: string
      - description: Container name (optional, defaults to first container)
        in: query
        name: container
        type: string
      - description: 'Number of lines to retrieve from the end (default: 1000, max:
          10000)'
        in: query
        name: tailLines
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: Pod logs in plain text
          schema:
            type: string
        "400":
          description: Invalid request parameters
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Pod not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Kubernetes API error
          schema:
            additionalProperties: true
            type: object
      summary: Get pod logs
      tags:
      - pods
...
  /api/v1/platform/pods:
    get:
      consumes:
      - application/json
      description: Lists all pods in the specified namespace
      parameters:
      - description: 'Kubernetes namespace (default: default)'
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of pods
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "502":
          description: Kubernetes API error
          schema:
            additionalProperties: true
            type: object
      summary: List pods
      tags:
      - pods
  /api/v1/platform/pods/{namespace}/{name}:
    delete:
      consumes:
      - application/json
      description: Deletes a specific pod
      parameters:
      - description: Kubernetes namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the pod
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deletion confirmation
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Pod not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Kubernetes API error
          schema:
            additionalProperties: true
            type: object
      summary: Delete pod
      tags:
      - pods
    get:
      consumes:
      - application/json
      description: Retrieves details of a specific pod
      parameters:
      - description: Kubernetes namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the pod
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Pod details
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Pod not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Kubernetes API error
          schema:
            additionalProperties: true
            type: object
      summary: Get pod
      tags:
      - pods
  /api/v1/platform/pods/{namespace}/{name}/exec:
    get:
      consumes:
      - application/json
      description: Establishes a WebSocket connection to execute commands in a pod's
        container
      parameters:
      - description: Kubernetes namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the pod
        in: path
        name: name
        required: true
        type: string
      - description: Container name (optional)
        in: query
        name: container
        type: string
      - description: 'Command to execute (default: /bin/sh)'
        in: query
        name: command
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: Switching protocols to WebSocket
          schema:
            type: string
        "400":
          description: Invalid request or WebSocket upgrade failed
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Pod not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Kubernetes API error
          schema:
            additionalProperties: true
            type: object
      summary: Execute command in pod via WebSocket
      tags:
      - pods
...
  /api/v1/polardbxclusters/{namespace}/{name}/pods:
    get:
      description: List pods belonging to the specified PolarDB-X cluster.
      parameters:
      - description: Kubernetes namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Cluster name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of pods
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "404":
          description: Cluster or pods not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Kubernetes API error
          schema:
            additionalProperties: true
            type: object
      summary: List cluster pods
      tags:
      - polardbxclusters
      - pods
...
  /api/v1/xstores/{namespace}/{name}/pods:
    get:
      description: List Pods under the specified XStore.
      parameters:
      - description: Kubernetes namespace of the XStore
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the XStore
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of Pods
          schema:
            items:
              $ref: '#/definitions/xstores.XStorePodDTO'
            type: array
        "404":
          description: XStore or Pods not found
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Upstream Kubernetes error
          schema:
            additionalProperties: true
            type: object
      summary: List XStore pods
      tags:
      - xstores
...
swagger: "2.0"